feat:no need to set ip and password for sealer run
