+++ utils/utils.go
@@ -33,10 +33,6 @@ func Retry(tryTimes int, trySleepTime time.Duration, action func() error) error 
 	return fmt.Errorf("retry action timeout: %v", err)
 }
 
-func WrapExecResult(host, command string, output []byte, err error) error {
-	return fmt.Errorf("failed to execute command(%s) on host(%s): output(%s), error(%v)", command, host, output, err)
-}
-
 // ConfirmOperation confirm whether to continue with the operation，typing yes will return true.
 func ConfirmOperation(promptInfo string) (bool, error) {
 	var yesRx = regexp.MustCompile("^(?:y(?:es)?)$")
